move join into its own file
